Timeout prevents sensing far objects and eliminates false near readings
Because during sumo competition the robot only needs to sense other robots in the ring, which is 77 cm in diameter, we can tell pulseIn to stop after a given number of microseconds -- in  this case the timeout is set to 7000 which in theory would correspond to 120 cm.  In practice robot's stop sensing and simply report 0's around 110 cm.  The ADVANTAGE of doing this is that it greatly reduces the chance of getting a false reading of a value near enough to be your opponent.  This false positive can be caused by a previous ultrasonic chirp bouncing back to the HC-SR04 after the robot has initiated but not yet finished a new ultrasound measurement.  The DISADVANTAGE is that in competition you will get many 0 distance values.  HOWEVER, this can easily be handled in your competition code, just look for values within the expected range, or reduce the timeout to whatever threshold value you like and then attack if a distance greater than 0 is read.