Including all TORTIS commands, a Logo subset
Some research revealed that there was a subset of Logo commands used in a pushblock format that enabled young children (unable to read and type yet) to use the robotic turlte (Primo.io's cubetto seems to have taken this idea and run with it).   Though "pen up" and "pen down" will require the addition of a servo motor and pen, they are included, though there implementation will be delayed until servos and engineering a pen-holder are accomplished.